requirements.txt
wxpython
python-vlc
yt-dlp
youtube-search-python
httpx==0.27.0
accessible-output2
passwordmeter
pwnedpasswords
pyOpenSSL
requests
pydub
keyboard
speechRecognition
pyautogui
pyperclip
paramiko
keyring
cryptography
win11toast
speedtest-cli
pyaudio
youtube_comment_downloader
srt